+++ lib/libraries/auth/data/models/auth_user.dart
@@ -1,36 +1,41 @@
 import 'package:construculator/libraries/auth/data/types/auth_types.dart';
 
 /// Represents a user in the application
-/// [id] is the unique identifier for the user, generated by the auth.users/credentials table
-/// [credentialId] is the uuid identifier for the user's credential, registered through the related auth.users/credentials table
-/// [email] is the email of the user
-/// [firstName] is the first name of the user
-/// [lastName] is the last name of the user
-/// [professionalRole] is a uuid that points to the professional_roles table, which contains the user's professional role
-/// [profilePhotoUrl] is the URL of the user's profile photo
-/// [createdAt] is the date and time the user was created
-/// [updatedAt] is the date and time the user was last updated
-/// [userStatus] is the status of the user, currently active or inactive.
-/// [userPreferences] is a map of user preferences, currently empty.
-///
-/// [fullName] is the full name of the user, calculated from the first and last name
-/// [copyWith] is a method to create a copy of the user with modified properties
-/// [empty] is a method to create an empty user
-/// [toJson] is a method to convert the user to a JSON object
-/// [fromJson] is a method to create a user from a JSON object
-
 class User {
+  /// The unique identifier for the user, generated by the auth.users/credentials table
   final String id;
+
+  /// The unique identifier for the user's credential, generated by the auth.users/credentials table
   final String credentialId;
+
+  /// The email of the user
   final String email;
+
+  /// The phone number of the user
   final String? phone;
+
+  /// The first name of the user
   final String firstName;
+
+  /// The last name of the user
   final String lastName;
+
+  /// The professional role of the user
   final String professionalRole;
+
+  /// The URL of the user's profile photo
   final String? profilePhotoUrl;
+
+  /// The date and time the user was created
   final DateTime createdAt;
+
+  /// The date and time the user was last updated
   final DateTime updatedAt;
+
+  /// The status of the user, currently active or inactive.
   final UserProfileStatus userStatus;
+
+  /// A map of user preferences, currently empty.
   final Map<String, dynamic> userPreferences;
 
   User({
@@ -48,6 +53,7 @@ class User {
     required this.userPreferences,
   });
 
+  /// Used to create a user from a JSON object
   factory User.fromJson(Map<String, dynamic> json) {
     return User(
       id: json['id'] as String,
@@ -65,6 +71,7 @@ class User {
     );
   }
 
+  /// Used to convert the user to a JSON object
   Map<String, dynamic> toJson() {
     return {
       'id': id,
@@ -82,6 +89,7 @@ class User {
     };
   }
 
+  /// Used to create an empty user
   static User empty() {
     return User(
       id: '',
@@ -99,10 +107,10 @@ class User {
     );
   }
 
-  // Helper to get full name
+  /// The full name of the user, computed from the first and last name
   String get fullName => '$firstName $lastName';
   
-  // Copy with method for creating modified instances
+  /// Used to create a copy of the user with modified properties
   User copyWith({
     String? id,
     String? credentialId,
