+++ lib/libraries/auth/data/types/auth_types.dart
@@ -5,8 +5,16 @@
 /// [connectionError] is used when the connection to the server is lost
 enum AuthStatus { authenticated, unauthenticated, connectionError }
 
+/// The type of address an otp is sent to
+/// 
+/// [email] indicates the otp should be sent to an email
+/// [phone] indicates the otp should be sent to a phone number
 enum OtpReceiver { email, phone }
 
+/// The user profile status
+/// 
+/// [active] indicates the user is active and mostly interracts with the platform
+/// [inactive] indicates the user is inactive and likely left the platform
 enum UserProfileStatus { active, inactive }
 
 /// Postgres error codes
@@ -24,6 +32,7 @@ enum PostgresErrorCode {
   unknownError,
 }
 
+/// Maps a string error code received form postgress to application level type
 PostgresErrorCode convertToPostgresErrorCode(String? errorCode) {
   switch (errorCode) {
     case '23505':
@@ -63,16 +72,19 @@ enum AuthErrorType {
 }
 
 /// Generic result class for authentication operations
-///
-/// [data] is the data returned from the operation
-/// [errorMessage] is the error message returned from the operation
-/// [errorType] is the type of error returned from the operation
-/// [isSuccess] is a boolean indicating if the operation was successful
 class AuthResult<T> {
+  /// The data returned from the operation
   final T? data;
+
+  /// The error message returned from the operation
   final String? errorMessage;
+
+  /// The type of error returned from the operation
   final AuthErrorType? errorType;
+
+  /// Indicates if the operation was successful
   final bool isSuccess;
+
   AuthResult.success(this.data)
     : isSuccess = true,
       errorMessage = null,
