+++ test/mutations/libraries/auth/supabase_repository.xml
@@ -77,14 +77,7 @@
     </rules>
 
     <commands>
-        <command group="auth-repositories" expected-return="0">flutter test test/units/libraries/auth/repositories/supabase_auth_repository_auth_state_test.dart</command>
-        <command group="auth-repositories" expected-return="0">flutter test test/units/libraries/auth/repositories/supabase_auth_repository_email_check_test.dart</command>
-        <command group="auth-repositories" expected-return="0">flutter test test/units/libraries/auth/repositories/supabase_auth_repository_email_login_test.dart</command>
-        <command group="auth-repositories" expected-return="0">flutter test test/units/libraries/auth/repositories/supabase_auth_repository_error_handling_part1_test.dart</command>
-        <command group="auth-repositories" expected-return="0">flutter test test/units/libraries/auth/repositories/supabase_auth_repository_error_handling_part2_test.dart</command>
-        <command group="auth-repositories" expected-return="0">flutter test test/units/libraries/auth/repositories/supabase_auth_repository_otp_auth_test.dart</command>
-        <command group="auth-repositories" expected-return="0">flutter test test/units/libraries/auth/repositories/supabase_auth_repository_profile_management_test.dart</command>
-        <command group="auth-repositories" expected-return="0">flutter test test/units/libraries/auth/repositories/supabase_auth_repository_registration_test.dart</command>
+        <command group="auth-repositories" expected-return="0">flutter test test/units/libraries/auth/supabase_auth_repository_test.dart</command>
     </commands>
 
     <threshold failure="85">
